+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+## [0.2.1]
 
 ### Fixed
 
@@ -82,6 +84,7 @@ All notable changes to this project will be documented in this file.
 * Neovim user commands to trigger UI keyboard shortcuts.
 
 [unreleased]: https://github.com/mickael-menu/ShadowVim/compare/main...HEAD
+[0.2.1]: https://github.com/mickael-menu/ShadowVim/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/mickael-menu/ShadowVim/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/mickael-menu/ShadowVim/compare/0.1.0...0.1.1
 [0.1]: https://github.com/mickael-menu/ShadowVim/tree/0.1.0